requirements.txt
shapely>=2.0.0
pyproj>=3.6.0
geopandas>=0.14.0
numpy>=1.26.0
scikit-learn>=1.3.0
python-igraph>=0.11.0
SQLAlchemy>=2.0.0
GeoAlchemy2>=0.15.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
pmtiles>=3.4.0
